group nav stuff into a single hook instead of HOC
