--- spotlight-client/src/NarrativeLayout/withNarrativeParams.tsx
@@ -1,74 +1,0 @@
-// Recidiviz - a data platform for criminal justice reform
-// Copyright (C) 2021 Recidiviz, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-// =============================================================================
-
-import { navigate } from "@reach/router";
-import { observer } from "mobx-react-lite";
-import React, { useCallback } from "react";
-import getUrlForResource from "../routerUtils/getUrlForResource";
-import { useDataStore } from "../StoreProvider";
-
-export type InjectedProps = {
-  navigateToSection: (s: number) => void;
-  sectionNumber: number;
-};
-
-export const withNarrativeParams = <OriginalProps extends unknown>(
-  OriginalComponent: React.ComponentType<OriginalProps & InjectedProps>
-): React.ComponentType<OriginalProps> => {
-  const NarrativeParamsProvider = (props: OriginalProps) => {
-    const {
-      tenantStore: {
-        currentTenantId: tenantId,
-        currentNarrativeTypeId: narrativeTypeId,
-        currentSectionNumber: sectionNumber,
-      },
-    } = useDataStore();
-
-    const navigateToSection = useCallback(
-      (newSectionNumber: number) => {
-        if (!tenantId || !narrativeTypeId) return;
-
-        navigate(
-          getUrlForResource({
-            page: "narrative",
-            params: {
-              tenantId,
-              narrativeTypeId,
-              sectionNumber: newSectionNumber,
-            },
-          }),
-          { replace: true }
-        );
-      },
-      [narrativeTypeId, tenantId]
-    );
-
-    if (
-      tenantId === undefined ||
-      narrativeTypeId === undefined ||
-      sectionNumber === undefined
-    ) {
-      return null;
-    }
-
-    return (
-      <OriginalComponent {...props} {...{ navigateToSection, sectionNumber }} />
-    );
-  };
-
-  return observer(NarrativeParamsProvider);
-};
